+++ test/insertion.jl
@@ -1,4 +1,4 @@
-module TestPush
+module TestInsertion
 
 using DataFrames, Test, Logging, DataStructures
 const ≅ = isequal
@@ -1281,6 +1281,66 @@ end
     end
     @test occursin("Error adding value to column :a2", String(take!(buf)))
     @test df == refdf
+end
+
+@testset "Tables.AbstractRow insertion" begin
+    tab = Tables.table([1 2 3; 4 5 6; 9 10 11; 12 13 14], header=[:a, :b, :c])
+    rows = tab |> Tables.rows |> collect
+    df = DataFrame()
+    @test push!(df, rows[2]) == DataFrame(a=4, b=5, c=6)
+    @test pushfirst!(df, rows[1], promote=false, cols=:union) ==
+          DataFrame(a=[1, 4], b=[2, 5], c=[3, 6])
+    @test insert!(df, 2, rows[3], promote=true, cols=:intersect) ==
+        DataFrame(a=[1, 9, 4], b=[2, 10, 5], c=[3, 11, 6])
+    @test push!(df, rows[1], promote=true, cols=:orderequal) ==
+        DataFrame(a=[1, 9, 4, 1], b=[2, 10, 5, 2], c=[3, 11, 6, 3])
+    deleteat!(df, nrow(df))
+
+    df2 = DataFrame(d="x", a="y")
+    push!(df2, rows[1], cols=:union)
+    @test df2 ≅ DataFrame(d=["x", missing], a=["y", 1], b=[missing, 2], c=[missing, 3])
+
+    tab = Tables.table(Any[15 16.5], header=[:d, :c])
+    row = tab |> Tables.rows |> first
+    @test push!(df, row, cols=:union) ≅
+                DataFrame(a=[1, 9, 4, missing],
+                          b=[2, 10, 5, missing],
+                          c=[3.0, 11.0, 6.0, 16.5],
+                          d=[missing, missing, missing, 15])
+
+    tab = Tables.table(Any[21 22.5], header=[:x, :b])
+    row = tab |> Tables.rows |> first
+    @test pushfirst!(df, row, cols=:subset) ≅
+                     DataFrame(a=[missing, 1, 9, 4, missing],
+                               b=[22.5, 2.0, 10.0, 5.0, missing],
+                               c=[missing, 3.0, 11.0, 6.0, 16.5],
+                               d=[missing, missing, missing, missing, 15])
+
+    tab = Tables.table(["a" "b" "c" "d" "e"], header=[:a, :b, :c, :d, :e])
+    row = tab |> Tables.rows |> first
+
+    buf = IOBuffer()
+    sl = SimpleLogger(buf)
+    with_logger(sl) do
+        @test_throws MethodError insert!(df, 3, row, cols=:intersect)
+    end
+    @test occursin("Error adding value to column :a", String(take!(buf)))
+
+    @test insert!(df, 3, row, cols=:intersect, promote=true) ≅
+                  DataFrame(a=[missing, 1, "a", 9, 4, missing],
+                            b=[22.5, 2.0, "b", 10.0, 5.0, missing],
+                            c=[missing, 3.0, "c", 11.0, 6.0, 16.5],
+                            d=[missing, missing, "d", missing, missing, 15])
+    for i in [1, 2, 4, 8, 16, 32, 64, 100, 1000, 10000, 20_000, 100_000]
+        df = DataFrame()
+        mat = Any[a + 100 * b + (iseven(b) ? 0.5 : 0) for a in 1:2, b in 1:i]
+        tab = Tables.table(mat, header=Symbol.("x", 1:i))
+        for row in Tables.rows(tab)
+            push!(df, row)
+        end
+        @test eltype.(eachcol(df)) == [(isodd(i) ? Int : Float64) for i in 1:ncol(df)]
+        @test df == DataFrame(mat, :auto)
+    end
 end
 
 end # module