+++ NEWS.md
@@ -3,13 +3,31 @@
 ## New functionalities
 
 * Add `Iterators.partition` support
-   ([#3212](https://github.com/JuliaData/DataFrames.jl/pull/3212))
+  ([#3212](https://github.com/JuliaData/DataFrames.jl/pull/3212))
 * Add `allunique` and allow transformations in `cols` argument of `describe`
   and `nonunique` when working with `SubDataFrame`
   ([3232](https://github.com/JuliaData/DataFrames.jl/pull/3232))
+* Add support for `Tables.AbstractRow` for `push!`, `pushfirst!`, and `insert!`
+   ([#3245](https://github.com/JuliaData/DataFrames.jl/pull/3245))
 * Add support for `operator` keyword argument in `Cols`
   to take a set operation to apply to passed selectors (`union` by default)
   ([3224](https://github.com/JuliaData/DataFrames.jl/pull/3224))
+* Improve support for setting group order in `groupby`
+  ([3253](https://github.com/JuliaData/DataFrames.jl/pull/3253))
+* Joining functions now support `order` keyword argument allowing the user
+  to specify the order of the rows in the produced table
+  ([#3233](https://github.com/JuliaData/DataFrames.jl/pull/3233))
+
+## Bug fixes
+
+* passing very many data frames to `innerjoin` and `outerjoin`
+  does not lead to stack overflow
+  ([#3233](https://github.com/JuliaData/DataFrames.jl/pull/3233))
+* fixed incorrect handling of passing no conditions in `subset` and `subset!`
+  ([#3264](https://github.com/JuliaData/DataFrames.jl/pull/3264))
+* fixed error in fast aggregation in `sum` and `mean` of columns only having
+  `missing` values
+  ([#3268](https://github.com/JuliaData/DataFrames.jl/pull/3268))
 
 # DataFrames.jl v1.4.4 Patch Release Notes